improve file organization
